add : authorization and update

+++ src/article/article.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateArticleDto } from './dto/create.article.dto';
 
@@ -6,22 +6,23 @@ import { CreateArticleDto } from './dto/create.article.dto';
 export class ArticleService {
   constructor(private prisma: PrismaService) {}
 
-  async create(request: CreateArticleDto) {
+  async create(request: CreateArticleDto, user: any) {
+    if (user.role !== 'CONTRIBUTOR' && user.role !== 'ADMIN') {
+      throw new UnauthorizedException('Unauthorized');
+    }
+
     const category = request.category;
-    const userId = 1;
 
     const value = await this.prisma.category.findFirst({
       where: { name: category },
     });
 
     const kategoriId = value.id;
-    const cover = 'https://dobrak-pagar/cover.jpg';
 
     const article = await this.prisma.article.create({
       data: {
         ...request,
-        cover: cover,
-        author: { connect: { id: userId } },
+        author: { connect: { id: user.id } },
         category: { connect: { id: kategoriId } },
       },
       include: {
@@ -66,23 +67,23 @@ export class ArticleService {
     return { ...article, author: authorName, category: categoryName };
   }
 
-  async update(articleId: number, request: CreateArticleDto) {
+  async update(articleId: number, request: CreateArticleDto, user: any) {
+    if (user.role !== 'CONTRIBUTOR' && user.role !== 'ADMIN') {
+      throw new UnauthorizedException('Unauthorized');
+    }
     const category = request.category;
-    const userId = 1;
 
     const value = await this.prisma.category.findFirst({
       where: { name: category },
     });
 
     const kategoriId = value.id;
-    const cover = 'XXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
 
     const article = await this.prisma.article.update({
       where: { id: articleId },
       data: {
         ...request,
-        cover: cover,
-        author: { connect: { id: userId } },
+        author: { connect: { id: user.id } },
         category: { connect: { id: kategoriId } },
       },
       include: {
@@ -98,7 +99,11 @@ export class ArticleService {
     return result;
   }
 
-  async deleteArticle(articleId: number) {
+  async deleteArticle(articleId: number, user: any) {
+    if (user.role !== 'CONTRIBUTOR' && user.role !== 'ADMIN') {
+      throw new UnauthorizedException('Unauthorized');
+    }
+
     await this.prisma.article.delete({
       where: { id: articleId },
     });
